backend/requirements.txt
flask>=3.0
python-dotenv>=1.0
supabase>=2.0
